Basis vector, vector scaling and linear dependency
This is my learning notes from the lecture series Essence of linear algebra by 3Blue1Brown.

I tried to visualise the idea of vectors as much as possible so I can really grasp the concept. 

The function "plot_vectors" first started based off of the plot_vectors by Mr. Hadrien Jean but I added more features to it. When you pass in a single vector, the plot will show you supporting dashed lines to clarify the vector position. If multiple, it shows how the vector "moves" by scaling and vector multiplication. So it is helpful if you pass in vectors that are scaled and multiplied to see how they move around across the 2D vector space. 

Hope it helps :) 

---------------------------------------------------------------------------------------
References
3Blue1Brown (2016) Linear combinations, span, and basis vectors | Chapter 2, Essence of linear algebra. https://www.youtube.com/watch?v=k7RM-ot2NWY&list=PLZHQObOWTQDPD3MizzM2xVFitgF8hE_ab&index=2&t=311s&ab_channel=3Blue1Brown